backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.0
boto3==1.29.0
python-dotenv==1.0.0
runpod==1.5.1
llama-cpp-python==0.2.11 